[Refactor] AlbumScreen UiState 수정
