+++ app/src/main/java/com/and04/naturealbum/ui/model/AlbumData.kt
@@ -1,0 +1,9 @@
+package com.and04.naturealbum.ui.model
+
+import com.and04.naturealbum.data.room.Label
+import com.and04.naturealbum.data.room.PhotoDetail
+
+data class AlbumData(
+    val label: Label,
+    val photoDetails: PhotoDetail
+)
